requirements.txt
streamlit==1.51.0
tensorflow==2.20.0
ultralytics==8.3.228
opencv-python==4.12.0.88
pillow==11.3.0
numpy==2.2.6
importlib-resources
